refactor(logging): simplify statsd logging decorator
